FuseChat-3.0/FuseEval/requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.9.4
aiosignal==1.3.1
alpaca_eval==0.6
annotated-types==0.7.0
anthropic==0.25.1
antlr4-python3-runtime==4.11.1
anyio==4.3.0
appdirs==1.4.4
argcomplete==3.5.1
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.45.0
black==24.10.0
blessed==1.20.0
boto3==1.34.117
botocore==1.34.117
build==1.2.2.post1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpickle==3.1.0
cohere==5.5.4
colorama==0.4.6
compressed-tensors==0.6.0
crashtest==0.4.1
cryptography==43.0.3
datamodel-code-generator==0.26.2
DataProperty==1.0.1
datasets==3.1.0
deepspeed==0.16.1
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
dulwich==0.21.7
editdistance==0.8.1
einops==0.8.0
email_validator==2.2.0
eval_type_backport==0.2.0
evalplus==0.2.1
evaluate==0.4.3
exceptiongroup==1.2.0
fastapi==0.115.3
fastavro==1.9.4
fastjsonschema==2.20.0
filelock==3.13.4
fire==0.7.0
flake8==7.1.1
flash-attn==2.6.3
frozenlist==1.4.1
fschat==0.2.36
fsspec==2024.2.0
genson==1.3.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
gpustat==1.1.1
greenlet==3.1.1
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
hf_transfer==0.1.8
hjson==3.1.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.26.1
idna==3.7
immutabledict==4.2.0
importlib_metadata==8.5.0
inflect==5.6.2
iniconfig==2.0.0
installer==0.7.0
interegular==0.3.3
isort==5.13.2
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.3.1
langchain==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.143
lark==1.1.9
latex2mathml==3.77.0
latex2sympy2==1.9.0
Levenshtein==0.26.1
livebench==0.0.4
llvmlite==0.43.0
lm-format-enforcer==0.10.6
lxml==5.3.0
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==3.0.2
mbstrdecoder==1.1.3
mccabe==0.7.0
mdurl==0.1.2
mistral_common==1.4.4
mistralai==0.1.8
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.18
ninja==1.11.1.2
nltk==3.9.1
numba==0.60.0
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
openai==1.54.5
opencv-python-headless==4.10.0.84
orjson==3.10.0
outlines==0.0.46
packaging==24.0
pandas==2.2.2
partial-json-parser==0.2.1.1.post4
pathspec==0.12.1
pathvalidate==3.2.1
patsy==0.5.6
Pebble==5.0.7
peft==0.13.2
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.2
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
portalocker==2.10.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.1.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.6
pycodestyle==2.12.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.2
pyext==0.7
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.2
pyproject_hooks==1.2.0
pytablewriter==1.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
RapidFuzz==3.10.1
ray==2.38.0
referencing==0.35.1
regex==2024.9.11
reka-api==3.2.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rouge-score==0.1.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.1
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.3
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlitedict==2.1.0
starlette==0.41.0
stop-sequencer==1.2.3
svgwrite==1.4.3
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.6
tempdir==0.7.1
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.6.0
timeout-decorator==0.5.0
together==1.3.4
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
torch==2.4.0
torchvision==0.19.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.45.2
tree-sitter==0.23.2
tree-sitter-python==0.23.2
triton==3.0.0
trl==0.12.2
trove-classifiers==2024.10.21.16
typepy==1.3.2
typer==0.13.0
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.32.0
uvloop==0.19.0
virtualenv==20.27.0
vllm==0.6.3.post1
wandb==0.19.0
watchfiles==0.22.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==13.1
wget==3.2
word2number==1.1
xformers==0.0.27.post2
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.2
zstandard==0.23.0
